added CoinDetailFragment.kt & WelcomeFragment.kt & CoinHistoryFragment.kt & CoinListFragment.kt & something refactoring with useCase

+++ app/src/main/java/com/handroid/currencyconverter/data/network/dto/detailinfo/CoinInfoDto.kt
@@ -25,7 +25,7 @@ class CoinInfoDto @Inject constructor(
     val price: Double?,
     @SerializedName("LASTUPDATE")
     @Expose
-    val lastUpdate: Float?,
+    val lastUpdate: Long?,
     @SerializedName("MEDIAN")
     @Expose
     val median: Double?,
